requirements.txt
fuzzywuzzy==0.18.0
json5==0.12.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.1.9
langchain-openai==0.3.28
langchain-tavily==0.2.11
langchain-text-splitters==0.3.9
langgraph==0.6.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langmem==0.0.29
langsmith==0.4.12
Levenshtein==0.27.1
mplfinance==0.12.10b0
openai==1.99.1
orjson==3.11.1
pandas==2.3.1
plotly==6.2.0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
streamlit==1.48.0
yfinance==0.2.65
